requirements.txt
streamlit
openai
python-dotenv
newspaper3k
textblob
nltk
lxml[html_clean]
